% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aIn}
\alias{aIn}
\title{Fuzzy string-set containment}
\usage{
aIn(s1, s2, ...)
}
\arguments{
\item{s1}{vector of patterns}

\item{s2}{vector of strings to be searched}
}
\description{
Match one set of strings within another set of strings
}

