\name{helperfunctions-package}
\alias{helperfunctions-package}
\alias{helperfunctions}
\docType{package}
\title{
\packageTitle{helperfunctions}
}
\description{
\packageDescription{helperfunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{helperfunctions}
\packageIndices{helperfunctions}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{helperfunctions}

Maintainer: \packageMaintainer{helperfunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
