% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inquire}
\alias{inquire}
\title{Inquire User}
\usage{
inquire(yesNoQuestion, answer = NULL)
}
\arguments{
\item{yesNoQuestion}{...}

\item{answer}{set a default answer for testing}
}
\description{
Ask the user a yes/no question
}

