% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{insert.vals}
\alias{insert.vals}
\title{Insert values}
\usage{
insert.vals(vector, values, indices)
}
\arguments{
\item{vector}{vector to be extended}

\item{values}{values to add to vector}

\item{indices}{locations in vector to insert values}
}
\description{
Extend a vector in internal positions with given values
}

