% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mgrep}
\alias{mgrep}
\title{Multi-Grep}
\usage{
mgrep(patterns, x, ignore.case = FALSE, strict = 0, fuzzy = FALSE,
  value = FALSE, perl = FALSE)
}
\arguments{
\item{value}{boolean - return positions or values}
}
\description{
Grep multiple patterns in a single string
}

