% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mgrepl}
\alias{mgrepl}
\title{Multi-Grepl}
\usage{
mgrepl(patterns, x, ignore.case = FALSE, strict = 0, fuzzy = FALSE,
  perl = FALSE)
}
\arguments{
\item{patterns}{vector of strings to search for}

\item{x}{character string in which to search for patterns}

\item{ignore.case}{boolean - is case important}

\item{strict}{integer - how many patterns should match for a TRUE}

\item{fuzzy}{boolean - allow fuzzy matching}

\item{perl}{boolean - allow perl regex}
}
\description{
Grepl multiple patterns in a single string
}

