% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mgsub}
\alias{mgsub}
\title{Multiple Gsub}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{vector of patterns to replace}

\item{replacement}{vector of replacements, synced to pattern}

\item{x}{character string to modify}
}
\description{
Replace multiple patterns within a character string
}

