% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeColumn}
\alias{normalizeColumn}
\title{Normalize Column}
\usage{
normalizeColumn(df, colname)
}
\arguments{
\item{df}{data.frame containing the column}

\item{colname}{character string of the column's name}
}
\description{
Normalize a column within a dataframe
}

