% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RandDF}
\alias{RandDF}
\title{'Random' data frame}
\usage{
RandDF(rows, cols)
}
\arguments{
\item{rows}{- integer number of rows}

\item{cols}{- integer number of cols}
}
\description{
Create a random dataframe of uniformly sampled integers
less than or equal to the product of the dimensions given
}

