% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeParens}
\alias{removeParens}
\title{Remove Parentheses}
\usage{
removeParens(s)
}
\arguments{
\item{s}{character string to be culled of parentheses}
}
\description{
Remove any parenthetical content from a character string
}

