% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SegmentVec}
\alias{SegmentVec}
\title{Segment a vector}
\usage{
SegmentVec(vec, num.segs, ordered = TRUE)
}
\arguments{
\item{vec}{the vector to butcher}

\item{num.segs}{number of segments to partition into}

\item{ordered}{boolean - should the vector be ordered?}
}
\description{
Chop a vector up into equally sized parts
}

