% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stringContains}
\alias{stringContains}
\title{String containment}
\usage{
stringContains(s1, s2)
}
\arguments{
\item{s1}{character string}

\item{s2}{character string}
}
\description{
See if string2 contains all of string1 (only alpha numeric)
}

