% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stringIntersect}
\alias{stringIntersect}
\title{String Intersection}
\usage{
stringIntersect(s1, s2)
}
\arguments{
\item{s1}{the first string}

\item{s2}{the second string}
}
\description{
Find all the words in common between two strings
}

